code recfactoring

+++ lib/data/models/monitoring_data.dart
@@ -1,6 +1,6 @@
 class MonitoringData {
   final DateTime timestamp;
-  final int value;
+  final double value;
   MonitoringData({required this.value, required this.timestamp});
   factory MonitoringData.fromjson(Map<String, dynamic> json){
     return MonitoringData(value: json['value'], timestamp:  DateTime.parse(json['timestamp']));
